refact:move class into domain, view package
