+++ src/main/java/racingcar/domain/Observer.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface Observer<T> {
     void observe(T event);
