+++ src/main/java/racingcar/domain/Racing.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Racing {
     private MovementCondition movementCondition;
