+++ src/main/java/racingcar/domain/RacingSubject.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.HashSet;
 import java.util.Set;
