+++ src/main/java/racingcar/domain/Subject.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface Subject<T> {
     void addObserver(Observer<T> observer);
