+++ src/test/java/RacingCarTest/NoMoveCondition.java
@@ -1,6 +1,6 @@
 package RacingCarTest;
 
-import racingcar.MovementCondition;
+import racingcar.domain.MovementCondition;
 
 public class NoMoveCondition implements MovementCondition {
     @Override
